Done task 1281
Not sure if it is the best solution, but it is the one, that came to me in a moment I finished reading the task. Faced problems of me unknowing String.charAt() returns ASCII value of char even if it is an Integer. Hardcoded it doing '-48'. Gonna know from now